+++ Cesium3DTilesSelection/src/TileContentLoadInfo.h
@@ -6,10 +6,10 @@
 #include <Cesium3DTilesSelection/TileID.h>
 #include <Cesium3DTilesSelection/TileRefine.h>
 #include <Cesium3DTilesSelection/TilesetOptions.h>
+#include <Cesium3DTilesSelection/TilesetSharedAssetSystem.h>
 #include <CesiumAsync/AsyncSystem.h>
 #include <CesiumAsync/IAssetAccessor.h>
 #include <CesiumGeometry/Axis.h>
-#include <CesiumGltfReader/GltfSharedAssetSystem.h>
 
 #include <gsl/span>
 #include <spdlog/fwd.h>
@@ -25,8 +25,8 @@ struct TileContentLoadInfo {
       const std::shared_ptr<IPrepareRendererResources>&
           pPrepareRendererResources,
       const std::shared_ptr<spdlog::logger>& pLogger,
-      const CesiumUtility::IntrusivePointer<
-          CesiumGltfReader::GltfSharedAssetSystem> maybeAssetDepot,
+      const CesiumUtility::IntrusivePointer<TilesetSharedAssetSystem>&
+          pSharedAssetSystem,
       const TilesetContentOptions& contentOptions,
       const Tile& tile);
 
@@ -43,8 +43,7 @@ struct TileContentLoadInfo {
   BoundingVolume tileBoundingVolume;
 
   std::optional<BoundingVolume> tileContentBoundingVolume;
-  CesiumUtility::IntrusivePointer<CesiumGltfReader::GltfSharedAssetSystem>
-      pAssetDepot;
+  CesiumUtility::IntrusivePointer<TilesetSharedAssetSystem> pSharedAssetSystem;
 
   TileRefine tileRefine;