+++ Cesium3DTilesSelection/src/TilesetContentManager.cpp
@@ -49,15 +49,15 @@ struct ContentKindSetter {
 
   void operator()(CesiumGltf::Model&& model) {
     for (CesiumGltf::Image& image : model.images) {
-      if (!image.pCesium)
+      if (!image.pAsset)
         continue;
 
       // If the image size hasn't been overridden, store the pixelData
       // size now. We'll be adding this number to our total memory usage soon,
       // and remove it when the tile is later unloaded, and we must use
       // the same size in each case.
-      if (image.pCesium->sizeBytes < 0) {
-        image.pCesium->sizeBytes = int64_t(image.pCesium->pixelData.size());
+      if (image.pAsset->sizeBytes < 0) {
+        image.pAsset->sizeBytes = int64_t(image.pAsset->pixelData.size());
       }
     }
 
@@ -157,7 +157,8 @@ getTileBoundingRegionForUpsampling(const Tile& parent) {
           CesiumGeospatial::BoundingRegion(
               globeRectangle,
               details.boundingRegion.getMinimumHeight(),
-              details.boundingRegion.getMaximumHeight()),
+              details.boundingRegion.getMaximumHeight(),
+              getProjectionEllipsoid(projection)),
           center};
     }
   }
@@ -565,8 +566,8 @@ postProcessContentInWorkerThread(
       tileLoadInfo.contentOptions.ktx2TranscodeTargets;
   gltfOptions.applyTextureTransform =
       tileLoadInfo.contentOptions.applyTextureTransform;
-  if (tileLoadInfo.pAssetDepot) {
-    gltfOptions.pSharedAssets = tileLoadInfo.pAssetDepot;
+  if (tileLoadInfo.pSharedAssetSystem) {
+    gltfOptions.pSharedAssetSystem = tileLoadInfo.pSharedAssetSystem;
   }
 
   auto asyncSystem = tileLoadInfo.asyncSystem;
@@ -666,9 +667,7 @@ TilesetContentManager::TilesetContentManager(
       _tileLoadsInProgress{0},
       _loadedTilesCount{0},
       _tilesDataUsed{0},
-      _pSharedAssets(CesiumGltfReader::GltfSharedAssetSystem::getDefault(
-          CesiumGltfReader::AssetSystemOptions{
-              tilesetOptions.contentOptions.ktx2TranscodeTargets})),
+      _pSharedAssetSystem(externals.pSharedAssetSystem),
       _destructionCompletePromise{externals.asyncSystem.createPromise<void>()},
       _destructionCompleteFuture{
           this->_destructionCompletePromise.getFuture().share()},
@@ -698,9 +697,7 @@ TilesetContentManager::TilesetContentManager(
       _tileLoadsInProgress{0},
       _loadedTilesCount{0},
       _tilesDataUsed{0},
-      _pSharedAssets(CesiumGltfReader::GltfSharedAssetSystem::getDefault(
-          CesiumGltfReader::AssetSystemOptions{
-              tilesetOptions.contentOptions.ktx2TranscodeTargets})),
+      _pSharedAssetSystem(externals.pSharedAssetSystem),
       _destructionCompletePromise{externals.asyncSystem.createPromise<void>()},
       _destructionCompleteFuture{
           this->_destructionCompletePromise.getFuture().share()},
@@ -852,9 +849,7 @@ TilesetContentManager::TilesetContentManager(
       _tileLoadsInProgress{0},
       _loadedTilesCount{0},
       _tilesDataUsed{0},
-      _pSharedAssets(CesiumGltfReader::GltfSharedAssetSystem::getDefault(
-          CesiumGltfReader::AssetSystemOptions{
-              tilesetOptions.contentOptions.ktx2TranscodeTargets})),
+      _pSharedAssetSystem(externals.pSharedAssetSystem),
       _destructionCompletePromise{externals.asyncSystem.createPromise<void>()},
       _destructionCompleteFuture{
           this->_destructionCompletePromise.getFuture().share()},
@@ -1000,7 +995,7 @@ void TilesetContentManager::loadTileContent(
       this->_externals.pAssetAccessor,
       this->_externals.pPrepareRendererResources,
       this->_externals.pLogger,
-      this->_pSharedAssets,
+      this->_pSharedAssetSystem,
       tilesetOptions.contentOptions,
       tile};
 
@@ -1243,9 +1238,9 @@ TilesetContentManager::getTilesetCredits() const noexcept {
   return this->_tilesetCredits;
 }
 
-const CesiumUtility::IntrusivePointer<CesiumGltfReader::GltfSharedAssetSystem>&
+const CesiumUtility::IntrusivePointer<TilesetSharedAssetSystem>&
 TilesetContentManager::getSharedAssetSystem() const noexcept {
-  return this->_pSharedAssets;
+  return this->_pSharedAssetSystem;
 }
 
 int32_t TilesetContentManager::getNumberOfTilesLoading() const noexcept {