+++ CesiumGltf/include/CesiumGltf/ImageAsset.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "CesiumAsync/SharedAsset.h"
 #include "CesiumGltf/Ktx2TranscodeTargets.h"
 #include "CesiumGltf/Library.h"
+#include "CesiumUtility/SharedAsset.h"
 
 #include <cstddef>
 #include <cstdint>
@@ -27,11 +27,11 @@ struct CESIUMGLTF_API ImageAssetMipPosition {
 };
 
 /**
- * @brief Holds {@link Image} properties that are specific to the glTF loader
- * rather than part of the glTF spec.
+ * @brief A 2D image asset, including its pixel data. The image may have
+ * mipmaps, and it may be encoded in a GPU compression format.
  */
 struct CESIUMGLTF_API ImageAsset final
-    : public CesiumAsync::SharedAsset<ImageAsset> {
+    : public CesiumUtility::SharedAsset<ImageAsset> {
   /**
    * @brief The width of the image in pixels.
    */