+++ CesiumGltf/src/TextureView.cpp
@@ -41,8 +41,8 @@ TextureView::TextureView(
     return;
   }
 
-  this->_pImage = model.images[static_cast<size_t>(texture.source)].pCesium;
-  if (this->_pImage->width < 1 || this->_pImage->height < 1) {
+  this->_pImage = model.images[static_cast<size_t>(texture.source)].pAsset;
+  if (!this->_pImage || this->_pImage->width < 1 || this->_pImage->height < 1) {
     this->_textureViewStatus = TextureViewStatus::ErrorEmptyImage;
     return;
   }