+++ CesiumGltf/test/TestFeatureIdTextureView.cpp
@@ -75,9 +75,9 @@ TEST_CASE("Test FeatureIdTextureView on feature ID texture with empty image") {
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 0;
-  image.pCesium->height = 0;
+  image.pAsset.emplace();
+  image.pAsset->width = 0;
+  image.pAsset->height = 0;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -108,10 +108,10 @@ TEST_CASE("Test FeatureIdTextureView on feature ID texture with too many bytes "
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
-  image.pCesium->bytesPerChannel = 2;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
+  image.pAsset->bytesPerChannel = 2;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -144,9 +144,9 @@ TEST_CASE(
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -177,9 +177,9 @@ TEST_CASE(
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -210,9 +210,9 @@ TEST_CASE("Test FeatureIdTextureView on feature ID texture with out of range "
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -242,9 +242,9 @@ TEST_CASE("Test FeatureIdTextureView on valid feature ID texture") {
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -275,9 +275,9 @@ TEST_CASE("Test FeatureIdTextureView with applyKhrTextureTransformExtension = "
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -325,9 +325,9 @@ TEST_CASE("Test FeatureIdTextureView with applyKhrTextureTransformExtension = "
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -379,14 +379,14 @@ TEST_CASE("Test FeatureIdTextureView with makeImageCopy = true") {
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
-  image.pCesium->pixelData.resize(featureIDs.size());
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
+  image.pAsset->pixelData.resize(featureIDs.size());
   std::memcpy(
-      image.pCesium->pixelData.data(),
+      image.pAsset->pixelData.data(),
       featureIDs.data(),
       featureIDs.size());
 
@@ -412,14 +412,14 @@ TEST_CASE("Test FeatureIdTextureView with makeImageCopy = true") {
 
   // Clear the original image data.
   std::vector<std::byte> emptyData;
-  image.pCesium->pixelData.swap(emptyData);
+  image.pAsset->pixelData.swap(emptyData);
 
   const ImageAsset* pImage = view.getImage();
   REQUIRE(pImage);
-  REQUIRE(pImage->width == image.pCesium->width);
-  REQUIRE(pImage->height == image.pCesium->height);
-  REQUIRE(pImage->channels == image.pCesium->channels);
-  REQUIRE(pImage->bytesPerChannel == image.pCesium->bytesPerChannel);
+  REQUIRE(pImage->width == image.pAsset->width);
+  REQUIRE(pImage->height == image.pAsset->height);
+  REQUIRE(pImage->channels == image.pAsset->channels);
+  REQUIRE(pImage->bytesPerChannel == image.pAsset->bytesPerChannel);
   REQUIRE(pImage->pixelData.size() == featureIDs.size());
 }
 
@@ -432,9 +432,9 @@ TEST_CASE("Test getFeatureID on invalid feature ID texture view") {
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 1;
-  image.pCesium->height = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 1;
+  image.pAsset->height = 1;
 
   Texture& texture = model.textures.emplace_back();
   texture.sampler = 0;
@@ -467,14 +467,14 @@ TEST_CASE("Test getFeatureID on valid feature ID texture view") {
   std::vector<uint8_t> featureIDs{1, 2, 0, 7};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
-  image.pCesium->pixelData.resize(featureIDs.size());
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
+  image.pAsset->pixelData.resize(featureIDs.size());
   std::memcpy(
-      image.pCesium->pixelData.data(),
+      image.pAsset->pixelData.data(),
       featureIDs.data(),
       featureIDs.size());
 
@@ -513,14 +513,14 @@ TEST_CASE("Test getFeatureID on view with applyKhrTextureTransformExtension = "
   std::vector<uint8_t> featureIDs{1, 2, 0, 7};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
-  image.pCesium->pixelData.resize(featureIDs.size());
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
+  image.pAsset->pixelData.resize(featureIDs.size());
   std::memcpy(
-      image.pCesium->pixelData.data(),
+      image.pAsset->pixelData.data(),
       featureIDs.data(),
       featureIDs.size());
 
@@ -565,14 +565,14 @@ TEST_CASE(
   std::vector<uint8_t> featureIDs{1, 2, 0, 7};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
-  image.pCesium->pixelData.resize(featureIDs.size());
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
+  image.pAsset->pixelData.resize(featureIDs.size());
   std::memcpy(
-      image.pCesium->pixelData.data(),
+      image.pAsset->pixelData.data(),
       featureIDs.data(),
       featureIDs.size());
 
@@ -622,14 +622,14 @@ TEST_CASE("Test getFeatureId on view with makeImageCopy = true") {
   sampler.wrapT = Sampler::WrapT::CLAMP_TO_EDGE;
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
-  image.pCesium->pixelData.resize(featureIDs.size());
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
+  image.pAsset->pixelData.resize(featureIDs.size());
   std::memcpy(
-      image.pCesium->pixelData.data(),
+      image.pAsset->pixelData.data(),
       featureIDs.data(),
       featureIDs.size());
 
@@ -655,7 +655,7 @@ TEST_CASE("Test getFeatureId on view with makeImageCopy = true") {
 
   // Clear the original image data.
   std::vector<std::byte> emptyData;
-  image.pCesium->pixelData.swap(emptyData);
+  image.pAsset->pixelData.swap(emptyData);
 
   REQUIRE(view.getFeatureID(0, 0) == 1);
   REQUIRE(view.getFeatureID(1, 0) == 2);
@@ -674,14 +674,14 @@ TEST_CASE("Test getFeatureID rounds to nearest pixel") {
   std::vector<uint8_t> featureIDs{1, 2, 0, 7};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
-  image.pCesium->pixelData.resize(featureIDs.size());
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
+  image.pAsset->pixelData.resize(featureIDs.size());
   std::memcpy(
-      image.pCesium->pixelData.data(),
+      image.pAsset->pixelData.data(),
       featureIDs.data(),
       featureIDs.size());
 
@@ -719,13 +719,13 @@ TEST_CASE("Test getFeatureID clamps values") {
   std::vector<uint8_t> featureIDs{1, 2, 0, 7};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
 
-  auto& data = image.pCesium->pixelData;
+  auto& data = image.pAsset->pixelData;
   data.resize(featureIDs.size() * sizeof(uint8_t));
   std::memcpy(data.data(), featureIDs.data(), data.size());
 
@@ -763,13 +763,13 @@ TEST_CASE("Test getFeatureID handles multiple channels") {
   std::vector<uint16_t> featureIDs{260, 512, 8, 17};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 2;
-  image.pCesium->bytesPerChannel = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 2;
+  image.pAsset->bytesPerChannel = 1;
 
-  auto& data = image.pCesium->pixelData;
+  auto& data = image.pAsset->pixelData;
   data.resize(featureIDs.size() * sizeof(uint16_t));
   std::memcpy(data.data(), featureIDs.data(), data.size());
 
@@ -805,13 +805,13 @@ TEST_CASE("Check FeatureIdTextureView sampling with different wrap values") {
   std::vector<uint8_t> featureIDs{1, 2, 0, 7};
 
   Image& image = model.images.emplace_back();
-  image.pCesium.emplace();
-  image.pCesium->width = 2;
-  image.pCesium->height = 2;
-  image.pCesium->channels = 1;
-  image.pCesium->bytesPerChannel = 1;
+  image.pAsset.emplace();
+  image.pAsset->width = 2;
+  image.pAsset->height = 2;
+  image.pAsset->channels = 1;
+  image.pAsset->bytesPerChannel = 1;
 
-  auto& data = image.pCesium->pixelData;
+  auto& data = image.pAsset->pixelData;
   data.resize(featureIDs.size() * sizeof(uint8_t));
   std::memcpy(data.data(), featureIDs.data(), data.size());