+++ CesiumGltfReader/include/CesiumGltfReader/ImageDecoder.h
@@ -1,7 +1,8 @@
+#pragma once
+
 #include "CesiumGltf/ImageAsset.h"
 #include "CesiumGltfReader/Library.h"
 
-#include <CesiumAsync/SharedAssetDepot.h>
 #include <CesiumUtility/IntrusivePointer.h>
 
 #include <gsl/span>
@@ -13,8 +14,7 @@
 namespace CesiumGltfReader {
 
 /**
- * @brief The result of reading an image with
- * {@link ImageDecoder::readImage}.
+ * @brief The result of reading an image with {@link ImageDecoder::readImage}.
  */
 struct CESIUMGLTFREADER_API ImageReaderResult {
 
@@ -36,6 +36,9 @@ struct CESIUMGLTFREADER_API ImageReaderResult {
   std::vector<std::string> warnings;
 };
 
+/**
+ * @brief Contains methods for reading and manipulating images.
+ */
 class ImageDecoder {
 public:
   /**
@@ -65,26 +68,6 @@ public:
    */
   static std::optional<std::string>
   generateMipMaps(CesiumGltf::ImageAsset& image);
-};
-
-/**
- * Used to construct an ImageAsset.
- */
-struct ImageAssetFactory : CesiumAsync::AssetFactory<CesiumGltf::ImageAsset> {
-  ImageAssetFactory(
-      const CesiumGltf::Ktx2TranscodeTargets& ktx2TranscodeTargets_)
-      : ktx2TranscodeTargets(ktx2TranscodeTargets_) {}
-
-  CesiumUtility::IntrusivePointer<CesiumGltf::ImageAsset>
-  createFrom(const gsl::span<const gsl::byte>& data) const override {
-    CesiumGltfReader::ImageReaderResult imageResult =
-        ImageDecoder::readImage(data, this->ktx2TranscodeTargets);
-    // TODO: report warnings and errors!
-    return imageResult.pImage;
-  }
-
-private:
-  const CesiumGltf::Ktx2TranscodeTargets ktx2TranscodeTargets;
 };
 
 } // namespace CesiumGltfReader