+++ CesiumGltfReader/src/GltfSharedAssetSystem.cpp
@@ -1,21 +1,23 @@
-#include <CesiumGltf/ImageAsset.h>
-#include <CesiumGltf/Schema.h>
 #include <CesiumGltfReader/GltfSharedAssetSystem.h>
-#include <CesiumGltfReader/ImageDecoder.h>
-#include <CesiumGltfReader/ImageReader.h>
-#include <CesiumGltfReader/SchemaAssetFactory.h>
 
-namespace CesiumGltfReader {
+using namespace CesiumAsync;
+using namespace CesiumGltf;
+using namespace CesiumGltfReader;
+using namespace CesiumUtility;
 
 namespace {
 
-CesiumUtility::IntrusivePointer<GltfSharedAssetSystem>
-createDefault(const AssetSystemOptions& options) {
+CesiumUtility::IntrusivePointer<GltfSharedAssetSystem> createDefault() {
   CesiumUtility::IntrusivePointer<GltfSharedAssetSystem> p =
       new GltfSharedAssetSystem();
 
-  p->pImage.emplace(
-      std::make_unique<ImageAssetFactory>(options.ktx2TranscodeTargets));
+  p->pImage.emplace(std::function(
+      [](const AsyncSystem& asyncSystem,
+         const std::shared_ptr<IAssetAccessor>& pAssetAccessor,
+         const NetworkImageAssetDescriptor& key)
+          -> Future<ResultPointer<ImageAsset>> {
+        return key.load(asyncSystem, pAssetAccessor);
+      }));
 
   p->pExternalMetadataSchema.emplace(std::make_unique<SchemaAssetFactory>());
 
@@ -24,10 +26,12 @@ createDefault(const AssetSystemOptions& options) {
 
 } // namespace
 
+namespace CesiumGltfReader {
+
 /*static*/ CesiumUtility::IntrusivePointer<GltfSharedAssetSystem>
-GltfSharedAssetSystem::getDefault(const AssetSystemOptions& options) {
+GltfSharedAssetSystem::getDefault() {
   static CesiumUtility::IntrusivePointer<GltfSharedAssetSystem> pDefault =
-      createDefault(options);
+      createDefault();
   return pDefault;
 }