+++ CesiumRasterOverlays/test/TestAddRasterOverlayToGltf.cpp
@@ -142,7 +142,8 @@ TEST_CASE("Add raster overlay to glTF") {
                     geometricError,
                     16.0, // the Max SSE used to render the geometry
                     details->rasterOverlayProjections[0],
-                    details->rasterOverlayRectangles[0]);
+                    details->rasterOverlayRectangles[0],
+                    Ellipsoid::WGS84);
 
             // Get a raster overlay texture of the proper dimensions.
             IntrusivePointer<RasterOverlayTile> pRasterTile =
@@ -266,8 +267,8 @@ TEST_CASE("Add raster overlay to glTF") {
   const Model& gltfBack = *resultBack.model;
 
   REQUIRE(gltfBack.images.size() == 1);
-  REQUIRE(gltfBack.images[0].pCesium != nullptr);
-  CHECK(!gltfBack.images[0].pCesium->pixelData.empty());
+  REQUIRE(gltfBack.images[0].pAsset != nullptr);
+  CHECK(!gltfBack.images[0].pAsset->pixelData.empty());
 
   REQUIRE(!gltfBack.meshes.empty());
   REQUIRE(!gltfBack.meshes[0].primitives.empty());